Sign In Style
finished the style for the sign in page
