--- rockwell-automation-simulator/src/pages/inicio-de-sesin1.module.css
@@ -1,77 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  background-color: #f8f8f8;
-}
-
-.formContainer {
-  display: flex;
-  width: 800px;
-  background-color: white;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-}
-
-.formLeft {
-  background-color: #ff5a5f;
-  width: 40%;
-  border-radius: 10px 0 0 10px;
-}
-
-.formRight {
-  width: 60%;
-  padding: 40px;
-}
-
-.title {
-  font-size: 24px;
-  margin-bottom: 10px;
-}
-
-.description {
-  font-size: 16px;
-  margin-bottom: 20px;
-  color: #666;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-}
-
-.formGroup {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 15px;
-}
-
-.input {
-  width: 48%;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-}
-
-.inputFull {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  margin-bottom: 15px;
-}
-
-.button {
-  background-color: #ff5a5f;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.button:hover {
-  background-color: #e14b4f;
-}
