+++ rockwell-automation-simulator/src/pages/signIn.module.css
@@ -1,0 +1,139 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
+
+.container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #f8f8f8;
+  font-family: poppins;
+}
+
+.formContainer {
+  display: flex;
+  width: 100%;
+  height: 100vh;
+  background-color: white;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+}
+
+.nameContainer{
+
+  width: 50%;
+  margin-bottom: 0;
+  
+}
+
+  
+@keyframes moveBackground {
+  0% {
+      background-position: 0% 0%;
+  }
+  50% {
+      background-position: 100% 0%;
+  }
+  100% {
+      background-position: 0% 0%;
+  }
+}
+
+
+
+.formLeft {
+  width: 45%;
+  border-radius: 0 25px 25px 0;
+  height: 100vh;
+  background-image: url('/gradient2.jpeg');
+  background-size: 240% auto;
+  animation: moveBackground 120s linear infinite;
+}
+
+.formRight {
+  width: 65%;
+  padding: 5px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  font-size: 2em;
+  margin-bottom: 0;
+  font-weight: 400;
+}
+
+.description {
+  font-size: 16px;
+  margin-bottom: 0px;
+  color: #666;
+  width: 73%;
+  text-align: center;
+  margin-top: 5px;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  width: 70%;
+  margin-top: 3vh;
+
+}
+
+
+.formGroup {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 0px;
+  
+  
+}
+
+.input {
+  width: 92%;
+  padding: 10px;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  margin-top: 0;
+  font-family: poppins;
+  font-weight: 500;
+}
+
+.inputFull {
+  width: 98%;
+  height: 4vh;
+  padding: 5px;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+  margin-bottom: 5px;
+  font-family: poppins;
+  font-weight: 500;
+}
+
+.button {
+  background-color: #EB2C2B;
+  color: white;
+  padding: 10px 20px;
+  border: none;
+  border-radius:25px;
+  cursor: pointer;
+  font-size: 16px;
+  margin-top: 4vh;
+  width: 100%;
+  align-self: center;
+  font-family: Poppins;
+  transition: 0.4s;
+}
+
+.button:hover {
+  background-color: #ff1a21;
+  scale: 1.05;
+}
+.InputTitle{
+  line-height: 0;
+  font-size: 1em;
+  font-family: Poppins;
+  font-weight: 500;
+  width: 68%;
+}
