+++ rockwell-automation-simulator/src/styles/logIn.module.css
@@ -22,11 +22,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-
-
     width: 100%;
-
-    
   }
   
   .imageContainer {
